+++ studio/src/schemaTypes/index.ts
@@ -33,6 +33,7 @@ import {pricingMatrix} from './objects/pricingMatrix'
 import {pricingList} from './objects/pricingList'
 import {policyNotes} from './objects/policyNotes'
 import {featureGrid} from './objects/featureGrid'
+import {pricingCalculator} from './objects/pricingCalculator'
 import {settings} from './singletons/settings'
 import {link} from './objects/link'
 import {blockContent} from './objects/blockContent'
@@ -84,4 +85,5 @@ export const schemaTypes = [
   pricingList,
   policyNotes,
   featureGrid,
+  pricingCalculator,
 ]